@@ -6,80 +6,83 @@
     "ng": "node_modules/@angular/cli/bin/ng",
     "upgradeTesteditorOrgDependencies": "yarn upgrade -S @testeditor",
     "prestart": "yarn run upgradeTesteditorOrgDependencies",
-    "start": "ng serve -H 0.0.0.0 --preserve-symlinks",
+    "start": "ng serve --host 0.0.0.0",
     "prebuild": "yarn run upgradeTesteditorOrgDependencies && ng lint",
     "build": "ng build",
     "bareBuild": "ng build",
     "pretest": "yarn run upgradeTesteditorOrgDependencies",
-    "test": "ng test --preserve-symlinks",
-    "test:once": "ng test --single-run --preserve-symlinks",
+    "test": "ng test",
+    "test:once": "ng test --watch=false",
     "pretestCIChrome": "yarn run upgradeTesteditorOrgDependencies",
-    "testCIChrome": "ng test --browser ChromeHeadless --single-run --log-level debug --preserve-symlinks",
+    "testCIChrome": "ng test --browsers ChromeHeadless --watch=false",
     "pretestCIFirefox": "yarn run upgradeTesteditorOrgDependencies",
-    "testCIFirefox": "ng test --browser Firefox --single-run --log-level debug --preserve-symlinks",
+    "testCIFirefox": "ng test --browsers Firefox --watch=false",
     "lint": "ng lint",
     "pree2e": "webdriver-manager update --ignore_ssl --gecko false --standalone false",
-    "e2e": "ng e2e -wu false",
+    "e2e": "ng e2e --webdriver-update=false",
     "docker:build": "docker build -t testeditor/web:snapshot .",
     "docker:run": "docker run -p 4200:4200 testeditor/web:snapshot &",
     "postversion": "git push && git push --tags"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.10",
-    "@angular/common": "^5.2.10",
-    "@angular/compiler": "^5.2.10",
-    "@angular/core": "^5.2.10",
-    "@angular/forms": "^5.2.10",
-    "@angular/http": "^5.2.10",
-    "@angular/platform-browser": "^5.2.10",
-    "@angular/platform-browser-dynamic": "^5.2.10",
-    "@angular/router": "^5.2.10",
-    "@testeditor/messaging-service": "~1.4.0",
-    "@testeditor/test-navigator": "~0.1.0",
-    "@testeditor/testeditor-commons": "~0.5.0",
-    "@testeditor/testexec-details": "~0.3.0",
-    "@testeditor/testexec-navigator": "~0.4.0",
-    "@testeditor/teststep-selector": "~0.1.0",
-    "@testeditor/user-activity": "~0.0.0",
-    "ace-builds": "^1.3.3",
-    "angular-auth-oidc-client": "^4.1.0",
-    "angular-split": "1.0.0-rc.1",
-    "angular2-jwt": "^0.2.3",
-    "bootstrap": "^4.1.2",
+    "@angular/animations": "^7.2.10",
+    "@angular/common": "^7.2.10",
+    "@angular/compiler": "^7.2.10",
+    "@angular/core": "^7.2.10",
+    "@angular/forms": "^7.2.10",
+    "@angular/http": "^7.2.10",
+    "@angular/platform-browser": "^7.2.10",
+    "@angular/platform-browser-dynamic": "^7.2.10",
+    "@angular/router": "^7.2.10",
+    "@auth0/angular-jwt": "^2.1.0",
+    "@testeditor/messaging-service": "~1.5.0",
+    "@testeditor/test-navigator": "~0.2.0",
+    "@testeditor/testeditor-commons": "~0.6.0",
+    "@testeditor/testexec-details": "~0.4.0",
+    "@testeditor/testexec-navigator": "~0.5.0",
+    "@testeditor/teststep-selector": "~0.2.0",
+    "@testeditor/user-activity": "~0.1.0",
+    "ace-builds": "^1.4.3",
+    "angular-auth-oidc-client": "^9.0.3",
+    "angular-split": "2.0.1",
+    "bootstrap": "^4.3.1",
     "core-js": "^2.5.5",
     "font-awesome": "^4.7.0",
     "jquery": "^3.3.1",
     "ng4-loading-spinner": "~1.2.0",
     "ngx-bootstrap": "^2.0.4",
-    "popper.js": "^1.14.3",
+    "popper.js": "^1.14.7",
     "prophecy": "^0.0.4",
     "requirejs": "^2.3.3",
-    "rxjs": "^5.5.10",
-    "ts-clipboard": "^1.0.16",
-    "zone.js": "^0.8.26"
+    "rxjs": "^6.4.0",
+    "rxjs-compat": "^6.0.0-rc.0",
+    "ts-clipboard": "^1.0.17",
+    "tslib": "^1.9.0",
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.10",
-    "@angular/language-service": "^5.2.10",
-    "@types/jasmine": "2.8.6",
-    "@types/node": "^9.6.6",
-    "codelyzer": "~4.3.0",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "^7.3.6",
+    "@angular/compiler-cli": "^7.2.10",
+    "@angular/language-service": "^7.2.10",
+    "@types/jasmine": "3.3.12",
+    "@types/node": "^11.11.4",
+    "codelyzer": "~5.0.0-beta.1",
     "jasmine-core": "^3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.2",
+    "karma": "~4.0.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.4.2",
+    "karma-cli": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.0.5",
     "karma-firefox-launcher": "^1.0.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.0.0",
     "karma-junit-reporter": "^1.2.0",
-    "protractor": "~5.3.1",
+    "protractor": "~5.4.2",
     "ts-mockito": "^2.3.0",
-    "ts-node": "~6.0.1",
-    "tslint": "^5.9.1",
-    "typescript": "~2.8.3"
+    "ts-node": "~8.0.3",
+    "tslint": "5.11.0",
+    "typescript": "~3.2.4"
   }
 }
