@@ -1,7 +1,7 @@
 import { Injectable, Injector } from '@angular/core';
 import { OidcSecurityService } from 'angular-auth-oidc-client';
 import { HttpInterceptor, HttpRequest, HttpHandler, HttpEvent } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 @Injectable()
 export class AuthInterceptor implements HttpInterceptor {
